+++ src/scenes/LevelSelect.ts
@@ -141,7 +141,8 @@ export default class LevelSelect extends Phaser.Scene {
 
 	private UICam!: Phaser.Cameras.Scene2D.BaseCamera | any;
 
-	private levelsKey = ['intro', 'airborne', 'dive', 'uppercut', 'charge', 'parasol', 'pogo', 'grenade', 'test-map-chain', 'test-map-bomb-vision', 'level-2', 'level-3', 'level-4'];
+	private levelsKey = ['grenade', 'intro', 'airborne', 'dive', 'uppercut', 'charge', 'parasol', 'pogo', 'grenade', 'final'];
+	// private levelsKey = ['intro-1', 'intro-2', 'intro-3', 'intro-4', 'intro-5', 'intro-6', 'intro-7', 'intro-8', 'intro-9', 'intro-10'];
 
 	private gamepad:Phaser.Input.Gamepad.Gamepad | undefined;
 	private SelectKey!: Phaser.Input.Keyboard.Key;
@@ -165,7 +166,8 @@ export default class LevelSelect extends Phaser.Scene {
 			this.selectedLevel = this.registry.get('current-level-index');
 			console.log(this.selectedLevel);
 		}
-		this.levelText.setText(this.levelsKey[this.selectedLevel]);
+		this.levelText.setText
+			('Level ' + (this.selectedLevel + 1) + ' - ' + this.levelsKey[this.selectedLevel]);
 
 	// load level button event
 		this.levelBack.setInteractive();
@@ -236,7 +238,8 @@ export default class LevelSelect extends Phaser.Scene {
 		}
 
 	// update visual
-		this.levelText.setText(this.levelsKey[this.selectedLevel]);
+		this.levelText.setText
+			('Level ' + (this.selectedLevel + 1) + ' - ' + this.levelsKey[this.selectedLevel]);
 	}
 
 	loadLevel()